requirements.txt
pandas>=2.0
requests>=2.28
pybaseball>=2.2
tqdm>=4.65
Pillow>=10.0
